miner/requirements.txt
PyGithub
redis
javalang